AAs should load from character data now, though will be fucked up if you already had stacked aas.

+++ common/eq_packet_structs.h
@@ -4243,7 +4243,7 @@ struct AARankInfo_Struct
 	int32 total_cost;
 	int32 expansion;
 	int32 category;
-	uint8 charges;
+	uint32 charges;
 	uint8 grant_only;
 	uint32 total_effects;
 	uint32 total_prereqs;
